feat(user-resume-integration) : 로컬 개발 환경 S3 mock 및 webhook 중간 상태 처리
- S3Uploader: mock-upload 모드 추가 — 로컬 임시 디렉토리에 파일 저장 후 localhost URL 반환
- MockFileController: @Profile("local") 로컬 전용 파일 서빙 컨트롤러 추가
- SecurityConfig: /mock-files/** permitAll 추가
- application-local.yml: aws.s3.mock-upload, websocket.allowed-origins 설정 추가
- ResumeServiceImpl: PENDING/ANALYZING 중간 상태 webhook 수신 시 조기 반환 처리

+++ db/init.sql
@@ -1096,7 +1096,7 @@ COMMENT ON COLUMN ai_usage_logs.created_at      IS '사용 기록 일시';
 CREATE TABLE ai_ops_settings (
     ai_ops_setting_id  BIGINT      NOT NULL DEFAULT 1,
     selected_model_id  BIGINT      NOT NULL,
-    monthly_budget     INTEGER     NOT NULL,
+    monthly_budget     NUMERIC(15,2) NOT NULL,
     alert_enabled      BOOLEAN     NOT NULL DEFAULT TRUE,
     alert_channel      VARCHAR(20) NOT NULL DEFAULT 'DISCORD',
     alert_threshold    INTEGER     NOT NULL DEFAULT 85,
